feat(CC-81): add create clinic e2e test and refactor folder organization
